Separate slash and chat commands
