+++ Src/POI.DiscordDotNet/Commands/ChatCommands/BeatSaber/BaseSongCommand.cs
@@ -3,7 +3,7 @@ using DSharpPlus.Entities;
 using ImageMagick;
 using Microsoft.Extensions.Logging;
 using NodaTime;
-using POI.DiscordDotNet.Commands.Modules.ChatCommands;
+using POI.DiscordDotNet.Commands.ChatCommands.Modules;
 using POI.DiscordDotNet.Extensions;
 using POI.Persistence.Repositories;
 using POI.ThirdParty.BeatSaver.Extensions;
@@ -17,7 +17,7 @@ using POI.ThirdParty.ScoreSaber.Models.Scores;
 using POI.ThirdParty.ScoreSaber.Models.Wrappers;
 using POI.ThirdParty.ScoreSaber.Services;
 
-namespace POI.DiscordDotNet.Commands.BeatSaber
+namespace POI.DiscordDotNet.Commands.ChatCommands.BeatSaber
 {
 	public abstract class BaseSongCommand : BeatSaberCommandsModule
 	{
@@ -352,7 +352,6 @@ namespace POI.DiscordDotNet.Commands.BeatSaber
 			await ctx.Message
 				.RespondAsync(messageBuilder)
 				.ConfigureAwait(false);
-
 
 			// Getting data for the second image
 			var beatSaviorProfileData = await BeatSaviorApiService.FetchBeatSaviorPlayerData(scoreSaberId).ConfigureAwait(false);
