+++ Src/POI.DiscordDotNet/Commands/ChatCommands/Modules/AdminCommandsModule.cs
@@ -1,6 +1,6 @@
 using DSharpPlus.CommandsNext;
 
-namespace POI.DiscordDotNet.Commands.Modules.ChatCommands
+namespace POI.DiscordDotNet.Commands.ChatCommands.Modules
 {
 	public class AdminCommandsModule : BaseCommandModule
 	{
