+++ Src/POI.DiscordDotNet/Commands/ChatCommands/Modules/BeatSaberCommandsModule.cs
@@ -1,6 +1,6 @@
 using DSharpPlus.CommandsNext;
 
-namespace POI.DiscordDotNet.Commands.Modules.ChatCommands
+namespace POI.DiscordDotNet.Commands.ChatCommands.Modules
 {
 	public class BeatSaberCommandsModule : BaseCommandModule
 	{
