--- Src/POI.DiscordDotNet/Commands/Modules/SlashCommands/BeatSaberSlashCommandsModule.cs
@@ -1,8 +1,0 @@
-﻿using DSharpPlus.SlashCommands;
-
-namespace POI.DiscordDotNet.Commands.Modules.SlashCommands
-{
-	public class BeatSaberSlashCommandsModule : ApplicationCommandModule
-	{
-	}
-}
