+++ Src/POI.DiscordDotNet/Commands/SlashCommands/Profile/ProfileSlashCommandsModule.cs
@@ -1,7 +1,7 @@
 ﻿using DSharpPlus.SlashCommands;
 using JetBrains.Annotations;
 
-namespace POI.DiscordDotNet.Commands.Profile
+namespace POI.DiscordDotNet.Commands.SlashCommands.Profile
 {
 	[SlashCommandGroup("profile", "Personal profile related commands"), UsedImplicitly]
 	public partial class ProfileSlashCommandsModule : ApplicationCommandModule
