+++ Src/POI.DiscordDotNet/Commands/SlashCommands/Test/TestSlashCommandsModule.cs
@@ -1,6 +1,6 @@
 using DSharpPlus.SlashCommands;
 
-namespace POI.DiscordDotNet.Commands.Test
+namespace POI.DiscordDotNet.Commands.SlashCommands.Test
 {
 	public class TestSlashCommandsModule : ApplicationCommandModule
 	{
