+++ Src/POI.DiscordDotNet/Commands/SlashCommands/Utils/UtilSlashCommandsModule.cs
@@ -1,6 +1,6 @@
 ﻿using DSharpPlus.SlashCommands;
 
-namespace POI.DiscordDotNet.Commands.Utils
+namespace POI.DiscordDotNet.Commands.SlashCommands.Utils
 {
 	public class UtilSlashCommandsModule : ApplicationCommandModule
 	{
